data_pipeline/src/requirements.txt
azure-functions==1.20.0
openai==1.68.2
azure-ai-textanalytics==5.3.0
azure-ai-formrecognizer==3.3.3
azure-search-documents==11.4.0b11
azure-storage-blob==12.22.0
azure-storage-queue==12.11.0
langchain==0.1.14
langchain-openai==0.1.1
pymupdf==1.24.1
beautifulsoup4==4.12.3
python-pptx==0.6.23
openpyxl==3.1.2
lxml==5.2.1
neologdn==0.5.2
omegaconf==2.3.0
python-dotenv==1.0.1
pandas==2.2.2
botbuilder-integration-aiohttp==4.16.2